fixing model errors

+++ AngularJSAuthentication.API/Models/TemplateTask.cs
@@ -23,12 +23,12 @@ namespace AngularJSAuthentication.API.Models
             Active = true;
         }
 
-        public int TemplateId { get; set; }
-        public int TaskId { get; set; }
+        public long TemplateId { get; set; }
+        public long TaskId { get; set; }
         public string Stage { get; set; }
         public string Environment { get; set; }
         public string Domain { get; set; }
-        public int Tminus { get; set; }
+        public long Tminus { get; set; }
 
         public virtual ICollection<ActivityTemplate> ActivityTemplates { get; set; }
         public virtual ICollection<ActivityTask> ActivityTasks { get; set; }
